Fatigue is now stored on the server as an integer
Resolves #66

+++ server/src/com/rscdaemon/scripting/AbstractScript.java
@@ -557,7 +557,7 @@ public abstract class AbstractScript
 	}
 	
 	@Override
-	public final void IncreaseXP(Skill skill, float amount)
+	public final void IncreaseXP(Skill skill, int amount)
 	{
 		scope.addLast(new IncreaseXPEvent(this, skill, amount));
 	}
